+++ python/love/commander/salinfo.py
@@ -1,8 +1,29 @@
+# This file is part of LOVE-commander.
+#
+# Copyright (c) 2023 Inria Chile.
+#
+# Developed by Inria Chile.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or at
+# your option any later version.
+#
+# This program is distributed in the hope that it will be useful,but
+# WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+# or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
+# for more details.
+#
+# You should have received a copy of the GNU General Public License along with
+# this program. If not, see <http://www.gnu.org/licenses/>.
+
+
 """Define the SAL Info subapplication, which provides the endpoints to request
 info from SAL.
 """
 from aiohttp import web
 from lsst.ts import salobj
+from lsst.ts import xml
 
 
 def create_app(*args, **kwargs):
@@ -16,19 +37,20 @@ def create_app(*args, **kwargs):
     salinfo_app = web.Application()
 
     domain = salobj.Domain()
-    available_idl_files = list(domain.idl_dir.glob("**/*.idl"))
-    names = [
-        file.name.split("_",)[
-            -1
-        ].replace(".idl", "")
-        for file in available_idl_files
-    ]
+    available_component_names = xml.subsystems
     if kwargs.get("remotes_len_limit") is not None:
-        names = names[: kwargs.get("remotes_len_limit")]
-
+        available_component_names = available_component_names[
+            : kwargs.get("remotes_len_limit")
+        ]
     salinfos = {}
-    for name in names:
-        salinfos[name] = salobj.SalInfo(domain, name)
+    salinfo_filled = False
+
+    async def connect_to_salinfo_instances():
+        global salinfo_filled
+        async with salobj.Domain():
+            for name in available_component_names:
+                salinfos[name] = salobj.SalInfo(domain, name)
+        salinfo_filled = True
 
     async def get_metadata(request):
         """Handle get metadata requests.
@@ -50,6 +72,9 @@ def create_app(*args, **kwargs):
                     "xml_version": "<XML version in format x.x.x>"
                 }
         """
+        if not salinfo_filled:
+            await connect_to_salinfo_instances()
+
         results = {
             salinfos[name].metadata.name: {
                 "sal_version": salinfos[name].metadata.sal_version,
@@ -100,6 +125,9 @@ def create_app(*args, **kwargs):
                     }
         }
         """
+        if not salinfo_filled:
+            await connect_to_salinfo_instances()
+
         accepted_categories = ["telemetry", "event", "command"]
         categories = request.rel_url.query.get("categories", "").split("-")
         categories = [c for c in categories if c in accepted_categories]
@@ -232,6 +260,9 @@ def create_app(*args, **kwargs):
                     },
                 }
         """
+        if not salinfo_filled:
+            await connect_to_salinfo_instances()
+
         accepted_categories = ["telemetry", "event", "command"]
         categories = request.rel_url.query.get("categories", "").split("-")
         categories = [c for c in categories if c in accepted_categories]
@@ -252,7 +283,7 @@ def create_app(*args, **kwargs):
         salinfo_app : object
             The SAL Info application
         """
-        for name in names:
+        for name in available_component_names:
             await salinfos[name].close()
         await domain.close()